TEC-4768 cleanup
