+++ src/main/java/com/echobox/api/linkedin/types/objectype/MultiLocaleString.java
@@ -15,21 +15,37 @@
  * limitations under the License.
  */
 
-package com.echobox.api.linkedin.types;
+package com.echobox.api.linkedin.types.objectype;
 
 import com.echobox.api.linkedin.jsonmapper.LinkedIn;
 
 import lombok.Getter;
 
+import java.util.Map;
+
 /**
- * 
+ * Multi local string POJO
+ * @see <a href="https://developer.linkedin.com/docs/ref/v2/object-types#MultiLocaleString">
+ * MultiLocalString</a>
  * @author joanna
  *
  */
-public class ShareDistributionTarget {
+public class MultiLocaleString {
   
+  /**
+   * Maps a locale to a localized version of the string. Each key is a Locale record converted 
+   * to string format, with the language, country and variant separated by underscores. 
+   * Examples: 'en', 'de_DE', 'en_US_WIN', 'de_POSIX', 'fr_MAC'.
+   */
   @Getter
-  @LinkedIn("linkedInDistributionTarget")
-  private DistributionTarget distributionTarget;
+  @LinkedIn
+  private Map<String, String> localized;
+  
+  /**
+   * The preferred locale to use, based on standard rules
+   */
+  @Getter
+  @LinkedIn
+  private Locale preferredLocale;
 
 }
