+++ src/test/resources/com.echobox.api.linkedin.jsonmapper/auditStamp.json
@@ -1,0 +1,4 @@
+{
+  "actor": "urn:li:person:123ABC",
+  "time": 1332187798000
+}
