+++ src/test/resources/com.echobox.api.linkedin.jsonmapper/share.json
